+++ server/helpers/fileSystem.js
@@ -10,32 +10,30 @@ export const rootPath = join(dirname(fileURLToPath(import.meta.url)), "../");
 
 export async function ensureDirectoryExists(directoryPath) {
   try {
-      await fs.promises.access(directoryPath);
+    await fs.promises.access(directoryPath);
   }
-  catch (error)
-  {
-      await fs.promises.mkdir(directoryPath, { recursive:true });
+  catch (error) {
+    await fs.promises.mkdir(directoryPath, { recursive: true });
   }
 }
 
 export async function saveFileStream(directory, fileName, sourceStream) {
   let destinationPath = join(directory, fileName);
-  
+
   await ensureDirectoryExists(directory);
 
   let destinationStream = fs.createWriteStream(destinationPath);
   try {
     await pump(sourceStream, destinationStream);
   }
-  finally
-  {
+  finally {
     destinationStream.close();
   }
 }
 
 export async function saveFileContent(directory, fileName, content) {
   let destinationPath = join(directory, fileName);
-  
+
   await ensureDirectoryExists(destinationPath);
   await fs.promises.writeFile(destinationPath, data);
 }
