+++ server/helpers/images.js
